requirements.txt
# ── LLM Inference ──────────────────────────────────────────────────────────────
ollama>=0.3.0

# ── Config & Environment ───────────────────────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0.0
